stashing progress on GA for #5
