+++ main.py
@@ -1,65 +1,11 @@
 #!/usr/bin/env python3
 
 from args import parse_args
-from lp import get_pages, get_sections
+from lp import get_pages, get_segments
 from gematria import direct_translation, atbash, vigenere, rot, running_shift, ATBASH, RUNE_LOOKUP
 from math import  totients
-
-class TextRetrievalSpec:
-    def __init__(self, nums: list, mode=get_pages):
-        """ The specification for how to retrieve text from Liber Primus
-        @param num  the number of the part (page, segment, line, etc.)
-        @param mode a function reference to the retrieval method
-        """
-        self.nums = nums
-        self.mode = mode
-
-    def retrieve(self):
-        return self.mode(self.nums)
-
-class CryptoSpec:
-    def __init__(self, scheme, key=None, shift=0, lookup=RUNE_LOOKUP, skips=None, excludes=None):
-        """ The specification for the cryptography to be performed
-        @param scheme   The encryption scheme to use
-        @param key      The key for the cipher
-        @param shift    The shift value for lookups
-        @param lookup   A dict of <rune->tuple of (list[char/bigram], gematria value)
-        @param skips    A dict of <rune->occurence number> to use original unkeyed lookup, key index
-                        doesn't increment and is continued on the next character.
-        @param excludes A dict of <rune->occurence number> to use original unkeyed lookup, key index
-                        does increment
-        """
-        self.scheme = scheme
-        self.key = key
-        self.shift = shift
-        self.lookup = lookup
-        self.skips = skips
-        self.excludes = excludes
-
-class SolutionSpec:
-    def __init__(self, retrieval: TextRetrievalSpec, crypto: CryptoSpec, show_runes: bool = False):
-        self.retrieval = retrieval
-        self.crypto = crypto
-        self.show_runes = show_runes
-
-    def run(self):
-        """ Generic cradle to run decryptions """
-        for num, text in zip(self.retrieval.nums, self.retrieval.retrieve()):
-            print(f"=== PAGE {num} ===")
-            if self.show_runes:
-                print(text)
-                print("-----")
-            if self.crypto.key:
-                plaintext = self.crypto.scheme(text,
-                                               key=self.crypto.key,
-                                               shift=self.crypto.shift,
-                                               lookup=self.crypto.lookup,
-                                               skips=self.crypto.skips,
-                                               excludes=self.crypto.excludes)
-                print(plaintext)
-            else:
-                plaintext = self.crypto.scheme(text, lookup=self.crypto.lookup, shift=self.crypto.shift, skips=self.crypto.skips)
-                print(plaintext)
+from specs import SolutionSpec, CryptoSpec, TextRetrievalSpec
+from ga import GeneticAlgorithm
 
 SOLUTIONS = {
     0 : SolutionSpec(TextRetrievalSpec([0]), CryptoSpec(atbash, lookup=ATBASH)),
@@ -73,7 +19,7 @@ SOLUTIONS = {
     9 : SolutionSpec(TextRetrievalSpec([9]), CryptoSpec(direct_translation)),
     10: SolutionSpec(TextRetrievalSpec([10]), CryptoSpec(direct_translation)),
     11: SolutionSpec(TextRetrievalSpec([11]), CryptoSpec(direct_translation)),
-    12: SolutionSpec(TextRetrievalSpec([5], mode=get_sections), CryptoSpec(vigenere,
+    12: SolutionSpec(TextRetrievalSpec([5], mode=get_segments), CryptoSpec(vigenere,
                   key="FIRFUMFERENCE",
                   skips={'ᚠ': [2, 3]},
                   excludes={'ᛖ': [1, 2, 3, 5, 7],
@@ -102,6 +48,8 @@ def known():
     """ Decrypt known pages """
     for spec in SOLUTIONS.values():
         spec.run()
+    ga = GeneticAlgorithm(0.01, initial_pool=[SOLUTIONS[0], SOLUTIONS[1]])
+    ga.evolve()
 
 def attempt_target():
     """ Attempt decryption on specified pieces of liber primus """
