+++ requirements.txt
@@ -1,0 +1,127 @@
+ansicolor==0.3.2
+argcomplete==3.2.3
+astroid==3.1.0
+autopep8==2.0.4
+Beaker==1.12.1
+beautifulsoup4==4.12.3
+binwalk==2.3.4
+black==24.4.1
+blivet==3.9.2
+blivet-gui==2.5.0
+boto3==1.34.78
+botocore==1.34.78
+Brlapi==0.8.5
+cffi==1.16.0
+chardet==5.2.0
+charset-normalizer==3.3.2
+click==8.1.7
+cryptography==41.0.7
+cupshelpers==1.0
+dasbus==1.7
+dbus-python==1.3.2
+dill==0.3.8
+distlib==0.3.7
+distro==1.9.0
+dnf==4.19.2
+docstring-to-markdown==0.15
+fb-re2==1.0.7
+fedora-third-party==0.10
+file-magic==0.4.0
+filelock==3.13.1
+flake8==7.0.0
+fros==1.1
+greenlet==3.0.3
+humanize==3.13.1
+idna==3.6
+importlib_metadata==7.1.0
+iso639==0.1.4
+isort==5.13.2
+jedi==0.19.1
+jmespath==1.0.1
+langtable==0.0.65
+libcomps==0.1.20
+libdnf==0.73.1
+lingua-language-detector==2.0.2
+louis==3.28.0
+lxml==5.1.0
+Mako==1.2.3
+MarkupSafe==2.1.3
+mccabe==0.7.0
+mercurial==6.7.2
+msgpack==1.0.8
+mypy==1.10.0
+mypy-extensions==1.0.0
+nftables==0.1
+olefile==0.47
+packaging==23.2
+parso==0.8.4
+Paste==3.7.1
+pathspec==0.12.1
+pexpect==4.9.0
+pid==2.2.3
+pillow==10.3.0
+platformdirs==4.2.1
+pluggy==1.5.0
+ply==3.11
+productmd==1.38
+ptyprocess==0.7.0
+pwquality==1.4.5
+pycairo==1.25.1
+pycodestyle==2.11.1
+pycparser==2.20
+pycrypto==2.6.1
+pycups==2.0.1
+pydocstyle==6.3.0
+pyenchant==3.2.2
+pyflakes==3.2.0
+PyGObject==3.48.2
+pykickstart==3.52
+pylint==3.1.0
+pylsp-mypy==0.6.8
+pynvim==0.5.0
+pyOpenSSL==23.2.0
+pyparted==3.13.0
+PySocks==1.7.1
+python-augeas==1.1.0
+python-dateutil==2.8.2
+python-lsp-black==2.0.0
+python-lsp-isort==0.1
+python-lsp-jsonrpc==1.1.2
+python-lsp-server==1.11.0
+python-meh==0.51
+python-pam==2.0.2
+pytoolconfig==1.3.1
+pyudev==0.24.1
+pyxdg==0.27
+PyYAML==6.0.1
+regex==2023.12.25
+requests==2.31.0
+requests-file==2.0.0
+requests-ftp==0.3.1
+rope==1.13.0
+rpm==4.19.1.1
+s3transfer==0.10.1
+selinux @ file:///builddir/build/BUILD/libselinux-3.6/src
+sepolicy @ file:///builddir/build/BUILD/selinux-3.6/python/sepolicy
+setools==4.4.4
+setuptools==69.0.3
+simpleaudio==1.0.4
+simpleline==1.9.0
+six==1.16.0
+snowballstemmer==2.2.0
+sos==4.7.1
+soupsieve==2.5
+systemd-python==235
+Tempita==0.5.2
+tomli==2.0.1
+tomlkit==0.12.4
+typing_extensions==4.11.0
+ujson==5.9.0
+urllib3==1.26.18
+vim-vint==0.3.21
+virtualenv==20.21.1
+whatthepatch==1.0.5
+xkbregistry==0.3
+yapf==0.40.2
+zipp==3.18.1
+zombie-imp==0.0.2
